UserService throwing exceptions & UserController /error GetMapping

+++ src/main/java/com/biava/crudthymeleaf/services/UserService.java
@@ -8,6 +8,8 @@ import org.springframework.stereotype.Service;
 
 import com.biava.crudthymeleaf.entities.User;
 import com.biava.crudthymeleaf.repositories.UserRepository;
+import com.biava.crudthymeleaf.services.exceptions.ResourceNotFoundException;
+import com.biava.crudthymeleaf.services.exceptions.UserAlreadyExistsException;
 
 @Service
 public class UserService {
@@ -20,19 +22,23 @@ public class UserService {
 	}
 	
 	public User insert(User user) {
-		return repository.save(user);
+		if(repository.findByEmail(user.getEmail()) == null) {
+			return repository.save(user);
+		}
+		throw new UserAlreadyExistsException("The email " + user.getEmail() + " is in use!");
 	}
-	
-	public void update(User user) {
-		repository.save(user);
-	}
-	
+		
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(repository.existsById(id)) {
+			repository.deleteById(id);
+		}
+		else {
+			throw new ResourceNotFoundException("Invalid user. The Id " + id + " is invalid!");
+		}
 	}
 	
 	public User findById(Long id) {
 		Optional<User> user = repository.findById(id);
-		return user.orElseThrow(() -> new RuntimeException());
+		return user.orElseThrow(() -> new ResourceNotFoundException("Invalid user. The Id " + id + " is invalid!"));
 	}
 }
